requirements.txt
Django>=5.0,<6.0
djangorestframework
django-cors-headers
gunicorn
uvicorn
whitenoise
dj-database-url
psycopg2-binary
